backend/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
sqlalchemy==2.0.36
pymysql==1.1.1
cryptography==44.0.0
python-dotenv==1.0.1
PyJWT==2.10.1
bcrypt==4.2.1
pydantic[email]==2.10.3
python-multipart==0.0.20
